+++ src/blocks/photoessay/photoessay.js
@@ -10,6 +10,7 @@ import './editor.scss';
 
 // Internal dependencies.
 import './photoessay-image';
+import blockIcons from '../../components/block-icons/';
 
 // WordPress dependencies.
 const {
@@ -53,7 +54,7 @@ const {
 registerBlockType( 'editorial/photoessay', {
 	title: __( 'Photo Essay' ),
 	description: __( 'Insert a row of photos with optional layouts.' ),
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="#c00" d="M19 7h-1V5h-4v2h-4V5H6v2H5c-1.1 0-2 .9-2 2v10h18V9c0-1.1-.9-2-2-2zm0 10H5V9h14v8z"></Path></SVG>,
+	icon: blockIcons('photoessay'),
 	category: 'bu-editorial',
 	attributes: {
 		layout: {