+++ src/blocks/stat/stats.js
@@ -14,6 +14,7 @@ import './editor.scss';
 // Internal dependencies.
 import './stat';
 import getAllowedFormats from '../../global/allowed-formats';
+import blockIcons from '../../components/block-icons/';
 
 // WordPress dependencies.
 const {
@@ -73,7 +74,7 @@ const getBlockClasses = ( className, stats ) => {
 registerBlockType( 'bu/stats', {
 	title: __( 'Stats' ),
 	description: __( 'Add one to four statistics to your content.' ),
-	icon: <SVG viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><Path fill="#c00" d="M19 7h-1V5h-4v2h-4V5H6v2H5c-1.1 0-2 .9-2 2v10h18V9c0-1.1-.9-2-2-2zm0 10H5V9h14v8z"></Path></SVG>,
+	icon: blockIcons('stat'),
 	category: 'bu',
 	attributes: {
 		align: {